pyproject.toml
[project]
name = "tsc-marketing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu>=1.10.0",
    "flask>=3.1.0",
    "gradio>=5.26.0",
    "langchain>=0.3.24",
    "line-bot-sdk>=3.17.0",
    "numpy>=2.2.5",
    "openai>=1.76.0",
    "sentence-transformers>=4.1.0",
    "torch>=2.7.0",
]
